changed a bunch of methods in BidsServiceHelper, which i refactored aswell from BidsServiceMethods
