+++ src/main/java/sidkbk/celemo/services/BidsServices.java
@@ -29,7 +29,7 @@ public class BidsServices {
     @Autowired
     UserRepository userRepository;
     @Autowired
-    BidsServiceMethods bidsServiceMethods;
+    BidsServiceHelper bidsServiceHelper;
 
 // Find all bids
     public List<Bids>findAllBids(){
@@ -45,9 +45,9 @@ public class BidsServices {
         // gets DTO, checks auction id from auction-repo
         Auction foundAuction = auctionRepository.findById(bidsDTO.getAuctionId())
                 .orElseThrow(() -> new RuntimeException("Auction does not exist!"));
-        bidsServiceMethods.checkFinished(foundAuction);
+        bidsServiceHelper.checkFinished(foundAuction);
         // auction owner check
-        bidsServiceMethods.checkAuctionOwner(foundAuction, foundUser);
+        bidsServiceHelper.checkAuctionOwner(foundAuction, foundUser);
         // makes new bid object
         Bids newBid = new Bids();
         // sets user found from DTO userID
@@ -55,12 +55,12 @@ public class BidsServices {
         newBid.setStartPrice(bidsDTO.getStartBid());
         newBid.setAuctionId(foundAuction.getId());
         // Checks if price is higher or lower than previous bid
-        newBid = BidsServiceMethods.bidMaxPriceCheck(bidsDTO, newBid);
+        newBid = BidsServiceHelper.bidMaxPriceCheck(bidsDTO, newBid);
         // Checks 3 things unfortunately,
         // 1. Check if startBid and maxBid is higher than auction startPrice,
         // 2. Checks if users balance is valid,
         // 3. Checks if users balance is less than starting bid
-        bidsServiceMethods.bidOkCheck(bidsDTO, foundAuction, foundUser);
+        bidsServiceHelper.bidOkCheck(bidsDTO, foundAuction, foundUser);
         // checks if auction has a bid
         if (foundAuction.isHasBids() && foundAuction.getBid() != null) {
             // checks if user has the same id as the owner of the auction. auction owner is not supposed to be able to bid on their own auction.
@@ -77,25 +77,21 @@ public class BidsServices {
                 } else {
 
                     // Switch check method compares Maxbid from both auctions, depending on who wins moves to the correct case.
-                    switch (bidsServiceMethods.bidWinCheck(auctionCurrentBid, newBid)) {
+                    switch (bidsServiceHelper.bidWinCheck(auctionCurrentBid, newBid)) {
                         // User Loses max bid.
                         case 1:
-                            bidsServiceMethods.userLoses(newBid, auctionCurrentBid, foundAuction);
-                            break;
-                        // User max bid Matches competing max bid
-                        case 2:
-                            bidsServiceMethods.userMatchesBid(newBid, auctionCurrentBid, foundAuction);
+                            bidsServiceHelper.userLoses(newBid, auctionCurrentBid, foundAuction);
                             break;
                         // User Wins.
-                        case 3:
-                            bidsServiceMethods.userWins(newBid, auctionCurrentBid, currentBidUser, foundAuction, foundUser);
+                        case 2:
+                            bidsServiceHelper.userWins(newBid, auctionCurrentBid, currentBidUser, foundAuction, foundUser);
                             break;
                         }
                 }
             }
         }
             // There are no previous bidders and the user has put a valid bid, wins automatically.
-            bidsServiceMethods.noPreviousBidsWin(foundUser, newBid, foundAuction);
+            bidsServiceHelper.noPreviousBidsWin(foundUser, newBid, foundAuction);
             return ResponseEntity.ok("Bid has been created, current price is " + foundAuction.getCurrentPrice());
     }
